jarvis-field-tech/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.147.0
pypdf>=5.0.0
requests>=2.32.0
